Dual language interface

+++ Install.sh
@@ -8,13 +8,35 @@ osascript -e "tell application \"Terminal\" to set normal text color of window 1
 
 clear
 
+loc=`locale | grep LANG | sed -e 's/.*LANG="\(.*\)_.*/\1/'`
+
+            if [ ! $loc = "ru" ]; then
+
+printf '\n\n*****   This program enables Continuity support for the Mac   ******\n'
+printf '*****   By installing Lilu.kext with plugins            ******\n'
+printf '*****   If you have installed the Bluetooth 4.0+ LE USB dongle    ******\n'
+printf '*****                        Version 1.6                          ******\n'
+
+                else
+
 printf '\n\n*****   Программное включение поддержки Continuity для макинтош   ******\n'
 printf '*****   C обновленным модулем Bluetooth до версии HCI 4.0+ c LE   ******\n'
 printf '*****   Посредством установки Lilu.kext с плагинами               ******\n'
-printf '*****                        Версия 1.5                          ******\n'
+printf '*****                        Версия 1.6                          ******\n'
+
+            fi
 
 sleep 0.5
 
+            if [ ! $loc = "ru" ]; then
+printf '\n    !!!   Your system '
+printf "`sw_vers -productName`"
+printf ': '; printf "`sw_vers -productVersion`" 
+printf '('
+printf "`sw_vers -buildVersion`"
+printf ') '
+printf '  !!!\n'
+                else
 printf '\n    !!!   Ваша система '
 printf "`sw_vers -productName`"
 printf ': '; printf "`sw_vers -productVersion`" 
@@ -22,7 +44,7 @@ printf '('
 printf "`sw_vers -buildVersion`"
 printf ') '
 printf '  !!!\n'
-
+            fi
 
 board=`ioreg -lp IOService | grep board-id | awk -F"<" '{print $2}' | cut -c 2- | rev | cut -c 3- | rev`
 
@@ -34,10 +56,15 @@ scontinuity=`defaults read /System/Library/Frameworks/IOBluetooth.framework/Vers
 
 continuity=`echo ${scontinuity//[^0-1]/}`
 
-
+            if [ ! $loc = "ru" ]; then
+printf '\n    !!!   This MAC board-id - '
+printf "$board"
+printf '   !!!\n\n'
+                else
 printf '\n    !!!   board-id этого макинтоша = '
 printf "$board"
 printf '   !!!\n\n'
+            fi
 
 legal=0
 case "$board" in
@@ -92,8 +119,8 @@ esac
 
 if [[ $legal = 0 ]] 
 	then 
-		sleep 3
-		clear && printf '\e[3J'
+            clear && printf '\e[3J'
+            if [ $loc = "ru" ]; then		
 		printf '\n    !!!   board-id этого макинтоша = '
 		printf "$board"
 		printf '   !!!\n\n'
@@ -101,18 +128,42 @@ if [[ $legal = 0 ]]
 		echo "Завершение программы. Выход"
 
         read -p "Для выхода нажмите любую клавишу" -n 1 -r
+                else
+        printf '\n    !!!   This MAC board-id = '
+        printf "$board"
+		printf '   !!!\n\n'
+        echo "For this mac the patch is impossible or not required." 
+		echo "End of program. Exit"
+        read -p "Press any key to close this window " -n 1 -r       
+            fi
         clear
         osascript -e 'tell application "Terminal" to close first window' & exit
 		
 		exit 
 fi
  
+            if [ ! $loc = "ru" ]; then	
+echo "Checking the status of system protection "
+                else
 echo "Проверка состояния системной защиты... "
+            fi
 csrset=`csrutil status | awk -F"status: " '{print $2}' | rev | cut -c 2- | rev`
 
 if [[ "$csrset" != "disabled" ]]
 	 then 
 		clear && printf '\e[3J'
+            if [ ! $loc = "ru" ]; then
+        printf '\n    !!!   System Integrity Protection enabled     !!!\n\n'
+		echo "    !!!    Cannot continue installation"
+		echo "    !!!    If protection is enabled the patch will not work."
+		echo "    !!!    To disable protection, boot into Recovery"
+        echo "    !!!    Or from installation media"
+		echo "    !!!    Run the terminal utility and execute"
+		echo "    !!!    csrutil disable command"
+		echo "    !!!    and after reboot, run this program again"
+		echo "    !!!    End of program. Exit"
+        read -p "Press any key to close this window " -n 1 -r
+		                else        
 		printf '\n    !!!   Защита целостности системы включена     !!!\n\n'
 		echo "    !!!    Продолжение установки невозможно"
 		echo "    !!!    Если защита включена патч не сработает"
@@ -122,12 +173,18 @@ if [[ "$csrset" != "disabled" ]]
 		echo "    !!!    и после перезагрузки запустите программу еще раз"
 		echo "    !!!    Завершение работы программы.Выход\n"
 		read -p "Для выхода нажмите любую клавишу" -n 1 -r
+            fi
         clear
         osascript -e 'tell application "Terminal" to close first window' & exit
 		exit 1
 fi
+
 printf '\n'
-printf '    !!! Защита целостности системы выключена\n\n'
+            if [ ! $loc = "ru" ]; then
+printf '    !!! System Integrity Protection Off - OK !!!\n\n'
+                else
+printf '    !!! Защита целостности системы выключена - OK !!!\n\n'
+            fi
 
 sleep 3
 printf '\e[3J'
@@ -137,25 +194,42 @@ kextset=0
 btframestat=0
 
 #printf '\nПроверяем необходимость установки поддержки Continuity\n'
+            if [ ! $loc = "ru" ]; then
+printf '\nGet information about the system\n\n'
+printf 'Bluetooth framework whitelist status ' 
+                else           
 printf '\nПолучаем информацию о системе\n\n'
 printf 'Состояние разрешения сценария Bluetooth для '
-printf "$board"
-printf ' \n'
-
+            fi
+printf "$board"' \n'
+            
+            
 if [ $continuity == 1 ]
 	then
 		btframestat=1
-		printf '\n    !!!         Сценарий Continuity разрешен         !!!\n'
+            if [ ! $loc = "ru" ]; then
+        printf '\n    !!!         Continuity support enabled         !!!\n'
+                else
+        printf '\n    !!!         Сценарий Continuity разрешен         !!!\n'
+            fi
 	else
+            if [ ! $loc = "ru" ]; then
+        printf '\n    !!! Continuity support disabled, patch required !!!\n'
+                else
 		printf '\n    !!! Сценарий Continuity запрещен, необходим патч !!!\n'
+            fi
 fi
-
-
 
 liluset=0
 arptset=0
 bt4leset=0
+
+            if [ ! $loc = "ru" ]; then
+printf '\nCheck for installed kernel extensions in /System/Library/Extensions/ \n\n'
+                else
 printf '\nПроверка установленных расширений ядра в /System/Library/Extensions/ \n\n'
+            fi
+
 if [  -f "/System/Library/Extensions/Lilu.kext/Contents/Info.plist" ]; then liluset=1; fi
 
 #printf '\nВаш liluset = '
@@ -175,11 +249,27 @@ if [  -f "/System/Library/Extensions/BT4LEContiunityFixup.kext/Contents/Info.pli
 #printf '\n\n'
 
 if [ $liluset == 1 ] || [ $arptset == 1 ] || [ $bt4leset == 1 ]; then kextset=1; fi
-if [[ $kextset = 0 ]]; then echo "    !!!   Расширения для поддержки Continuity не установлены"; echo
+
+            
+if [[ $kextset = 0 ]]; then
+            if [ ! $loc = "ru" ]; then
+    echo "    !!!   Kernel extensions to support Continuity not installed"; echo
+                else
+    echo "    !!!   Расширения для поддержки Continuity не установлены"; echo
+            fi
 	else
-	echo "    !!!         Расширения уже установлены.          !!! "
-	echo
+            if [ ! $loc = "ru" ]; then
+    echo "    !!!         Kernel extensions to support Continuity installed          !!! "; echo
+                else
+	echo "    !!!         Расширения уже установлены.          !!! "; echo
+
+            fi
+
+            if [ ! $loc = "ru" ]; then
+    echo "Check for kernel extension versions ..."
+                else
 	echo "Проверяем версии расширенией ..."
+            fi
 
 	sliluver=`plutil -p /System/Library/Extensions/Lilu.kext/Contents/Info.plist | grep CFBundleVersion | awk -F"=> " '{print $2}' | cut -c 2- | rev | cut -c 2- | rev`
 	sairportver=`plutil -p /System/Library/Extensions/AirportBrcmFixup.kext/Contents/Info.plist | grep CFBundleVersion | awk -F"=> " '{print $2}' | cut -c 2- | rev | cut -c 2- | rev`
@@ -190,71 +280,63 @@ cd $(dirname $0)
 	gairportver=`plutil -p Other/AirportBrcmFixup.kext/Contents/Info.plist | grep CFBundleVersion | awk -F"=> " '{print $2}' | cut -c 2- | rev | cut -c 2- | rev`
 	gbt4lever=`plutil -p Other/BT4LEContiunityFixup.kext/Contents/Info.plist | grep CFBundleVersion | awk -F"=> " '{print $2}' | cut -c 2- | rev | cut -c 2- | rev`
 
-printf '\n !!! Установлен Lilu.kext - '
-printf "$sliluver"
-printf ' , устанавливается - '
-printf "$gliluver"
-printf '\n'
 
-
-printf ' !!! Установлен AirportBrcmFixup.kext - '
-printf "$sairportver"
-printf ' , устанавливается - '
-printf "$gairportver"
-printf '\n'
-
-printf ' !!! Установлен BT4LEContiunityFixup.kext - '
-printf "$sbt4lever"
-printf ' , устанавливается - '
-printf "$gbt4lever"
-printf '\n\n'
-
+            if [ ! $loc = "ru" ]; then
+printf '\n !!! installed Lilu.kext - '"$sliluver"' , is being installed - '"$gliluver"'\n'
+printf ' !!! installed AirportBrcmFixup.kext - '"$sairportver"' , is being installed - '"$gairportver"'\n'
+printf ' !!! installed BT4LEContiunityFixup.kext - '"$sbt4lever"' , is being installed - '"$gbt4lever"'\n\n'
+                else
+printf '\n !!! Установлен Lilu.kext - '"$sliluver"' , устанавливается - '"$gliluver"'\n'
+printf ' !!! Установлен AirportBrcmFixup.kext - '"$sairportver"' , устанавливается - '"$gairportver"'\n'
+printf ' !!! Установлен BT4LEContiunityFixup.kext - '"$sbt4lever"' , устанавливается - '"$gbt4lever"'\n\n'
+            fi
 
 fi
-
-
-
-#liluver=0
-#arptver=0
-#bt4lever=0
-
-#printf '\nПроверка загружены ли расширения ядра в память  \n\n'
-#liluver=`kextstat | grep as.vit9696.Lilu | awk -F"as.vit9696.Lilu " '{print $2}' | cut -c 2- | rev | awk -F ")" '{print $2}' | rev`
-#arptver=`kextstat | grep as.lvs1974.BT4LEContiunityFixup  | awk -F"as.lvs1974.BT4LEContiunityFixup " '{print $2}' | cut -c 2- | rev | awk -F ")" '{print $2}' | rev`
-#bt4lever=`kextstat | grep as.lvs1974.AirportBrcmFixup  | awk -F"as.lvs1974.AirportBrcmFixup " '{print $2}' | cut -c 2- | rev | awk -F ")" '{print $2}' | rev`
 
 
 if [ $kextset == 1 ] && [ $btframestat == 1 ]
 	then
+            if [ ! $loc = "ru" ]; then
+        echo "It seems you already have Continuity support installed"
+		echo "Continue if you want to upgrade or reinstall the kernel extensions"
+                else
 		echo "Кажется у вас уже установлена поддержка Continuity"
 		echo "Продолжайте если хотите обновить или переустановить расширения"
+            fi
 fi
+            if [ ! $loc = "ru" ]; then
+        echo "To continue, press the English letter Y"
+		echo "To end the program any other key"
+        printf '\n\n'
+        read -p "Would you like to continue the installation? (y/N) " -n 1 -r
+                else
 		echo "Для продолжения нажмите англ. литеру Y"
 		echo "Для завершения любую другую клавишу"
-		printf '\n\n'
-               
-
-read -p "Желаете продолжить установку? (y/N) " -n 1 -r
+        printf '\n\n'
+        read -p "Желаете продолжить установку? (y/N) " -n 1 -r
+            fi
 
 if [[ ! $REPLY =~ ^[yY]$ ]]
-
 then
-    
+            if [ ! $loc = "ru" ]; then 
+    printf '\n\nWise choice. The End of the program. Exit ... !\n'
+                else   
     printf '\n\nМудрый выбор. Завершение  программы. Выход ... !\n'
+            fi
     sleep 3
-    
-        clear
+    clear
         osascript -e 'tell application "Terminal" to close first window' & exit
-   
-    exit 1
-
+   exit 
 fi
 
 sleep 0.3
-
+            if [ ! $loc = "ru" ]; then
+printf '\n\n\n*****  You know exactly what you are doing!  *****\n\n'
+printf '\nEnter your password to continue\n\n'
+                else
 printf '\n\n\n*****  Вы точно знаете что делаете!  *****\n\n'
 printf '\nДля продолжения введите ваш пароль\n\n'
-
+            fi
 
 
 sudo printf '\n\n'
@@ -271,7 +353,11 @@ let _left=40-$_done
 _fill=$(printf "%${_done}s")
 _empty=$(printf "%${_left}s")
 
+            if [ ! $loc = "ru" ]; then
+printf "\rRunning: ${_fill// /.}${_empty// / } ${_progress}%%"
+                else
 printf "\rВыполняется: ${_fill// /.}${_empty// / } ${_progress}%%"
+            fi
 
 }
 
@@ -359,12 +445,16 @@ sudo chmod -R 755 /System/Library/Extensions/BT4LEContiunityFixup.kext
 number=60
 ProgressBar ${number} ${_end}
 
+if [ ! -f "/System/Library/CoreServices/MRT.app.back/Contents/Info.plist" ]; then
 sudo cp -R /System/Library/CoreServices/MRT.app /System/Library/CoreServices/MRT.app.back
+fi
 
 number=65
 ProgressBar ${number} ${_end}
 
-sudo rm  -R -f /System/Library/CoreServices/MRT.app
+if [ -f "/System/Library/CoreServices/MRT.app/Contents/Info.plist" ]; then
+sudo rm  -R /System/Library/CoreServices/MRT.app
+fi
 
 number=70
 ProgressBar ${number} ${_end}
@@ -406,33 +496,48 @@ fi
 number=100
 ProgressBar ${number} ${_end}
 
+            if [ ! $loc = "ru" ]; then
+printf '\n\n.   !!!    Kernel Extensions Installed or Updated\n\n'
+                else
 printf '\n\n.   !!!    Расширения ядра установлены или обновлены\n\n'
-
+            fi
 if [[ $btframestat == 0 ]]
 	then
+            if [ ! $loc = "ru" ]; then
+        echo "    !!!    Bluetooth whitelist patch for Continuity is made"
+                else
 		echo "    !!!    Патч сценария Bluetooth для Continuity сделан"
+            fi
 fi
 
 sleep 1
 
-
-
+            if [ ! $loc = "ru" ]; then
+printf '\nWe update the system cache.\n'
+printf '\nProcessing: \n'
+                else
 printf '\nОбновляем системный кэш.\n'
 printf '\nВыполняется: \n'
+            fi
 
 while :;do printf '.\n' ;sleep 7;done &
 trap "kill $!" EXIT 
 sudo touch /System/Library/Extensions 2>/dev/null
 sudo kextcache -u /  2>/dev/null
+sleep 10
 kill $!
 wait $! 2>/dev/null
 trap " " EXIT
 
-
+            if [ ! $loc = "ru" ]; then
+printf '\n\nKernel kexts cache updated\n'
+sleep 1
+printf '\n\nTimeout for system setting\n\n'
+                else
 printf '\n\nСистемный кэш обновлен\n'
 sleep 1
-
 printf '\n\nТаймаут для системного урегулирования\n\n'
+            fi
 
 
 
@@ -450,29 +555,37 @@ done
 if [[ $btframestat == 0 ]]
 	then
 
+            if [ ! $loc = "ru" ]; then
+printf '\n\nupdate the system frameworks cache. It takes a few minutes\n'
+printf '\nProcessing: '
+                else
 printf '\n\nОбновляем кэш системных сценариев. Это занимает несколько минут\n'
+printf '\nВыполняется: '
+            fi
 sleep 1
 
 
-
-printf '\nВыполняется: '
 while :;do printf '.';sleep 3;done &
 trap "kill $!" EXIT 
  sudo update_dyld_shared_cache -debug -force -root / 2>/dev/null
+
 kill $!
 wait $! 2>/dev/null
 trap " " EXIT
 
+            if [ ! $loc = "ru" ]; then
+printf '\n\nSystem Frameworks Cache Updated\n\n'
+printf '\nAll operations completed\n'
+sleep 1
+printf '\nRequired timeout before rebooting the operating system\n'
+printf '\nPress CTRL + C to brake if you do not want to restart now\n\n'
+                else
 printf '\n\nКэш системных сценариев обновлен\n\n'
-
-
-
 printf '\nВсе операции завершены\n'
 sleep 1
-
 printf '\nНеобходимый таймаут перед перезагрузкой операционной системы\n'
-
 printf '\nНажмте CTRL + C для прерывания если не хотите перезагружать сейчас\n\n'
+            fi
 sleep 5
 
 
@@ -491,12 +604,14 @@ done
 sleep 1
 
 fi
-
+            if [ ! $loc = "ru" ]; then
+printf '\n\nThe program is complete. Initiated reboot. It may takes a couple of minutes.\n\n'
+                else
 printf '\n\nПрограмма завершена. Инициирована перезагрузка. На HDD может занять пару минут.\n\n'
-
+            fi
 
 sudo reboot now
 
-exit 1
+exit
